Remove explicit updating of well AD vars.
Now that the AD (Evaluation) primary variables in the well classes
are always updated whenever the value primary variables are, we
no longer need to remember to explicitly update them, simplifying
the code and enabling removal of functions.

The init() function of the primary variable classes have been renamed
to setEvaluationsFromValues() and made private.